# Improve NEON32 encoding speed