The enc_translate function now uses the same tricks as the SSSE3 one.

Speed-up on `iPhone SE` using `Apple LLVM version 8.0.0 (clang-800.0.38)`
NEON32: +33% compared to previous version

Full results before & after modifications follows.
Before:
```
Filling buffer with 10.0 MB of random data...
Testing with buffer size 10 MB, fastest of 100 * 1
NEON32	encode	2174.78 MB/sec
NEON32	decode	1408.59 MB/sec
plain	encode	1004.75 MB/sec
plain	decode	706.02 MB/sec
Testing with buffer size 1 MB, fastest of 100 * 10
NEON32	encode	2191.70 MB/sec
NEON32	decode	1413.05 MB/sec
plain	encode	1043.34 MB/sec
plain	decode	708.30 MB/sec
Testing with buffer size 100 KB, fastest of 100 * 100
NEON32	encode	2189.48 MB/sec
NEON32	decode	1411.65 MB/sec
plain	encode	1028.19 MB/sec
plain	decode	713.58 MB/sec
Testing with buffer size 10 KB, fastest of 1000 * 100
NEON32	encode	2205.67 MB/sec
NEON32	decode	1418.22 MB/sec
plain	encode	1053.66 MB/sec
plain	decode	719.94 MB/sec
Testing with buffer size 1 KB, fastest of 1000 * 1000
NEON32	encode	2057.55 MB/sec
NEON32	decode	1312.66 MB/sec
plain	encode	1021.82 MB/sec
plain	decode	705.21 MB/sec
```
After:
```
Filling buffer with 10.0 MB of random data...
Testing with buffer size 10 MB, fastest of 100 * 1
NEON32	encode	2940.38 MB/sec
NEON32	decode	1427.19 MB/sec
plain	encode	1055.70 MB/sec
plain	decode	721.25 MB/sec
Testing with buffer size 1 MB, fastest of 100 * 10
NEON32	encode	2943.92 MB/sec
NEON32	decode	1428.84 MB/sec
plain	encode	1054.56 MB/sec
plain	decode	722.47 MB/sec
Testing with buffer size 100 KB, fastest of 100 * 100
NEON32	encode	2937.40 MB/sec
NEON32	decode	1425.56 MB/sec
plain	encode	1052.89 MB/sec
plain	decode	722.08 MB/sec
Testing with buffer size 10 KB, fastest of 1000 * 100
NEON32	encode	2927.49 MB/sec
NEON32	decode	1418.48 MB/sec
plain	encode	1053.09 MB/sec
plain	decode	725.24 MB/sec
Testing with buffer size 1 KB, fastest of 1000 * 1000
NEON32	encode	2685.02 MB/sec
NEON32	decode	1334.63 MB/sec
plain	encode	1016.33 MB/sec
plain	decode	710.38 MB/sec
```