@mayeut 2016-10-16
A check on another device and/or with another compiler might be a good idea before merging this in.