@aklomp 2016-10-17
Tested on the Raspberry 2 with `Debian clang version 3.0-6.2` shows the same result:

```
Filling buffer with 10.0 MB of random data...
Testing with buffer size 10 MB, fastest of 10 * 1
NEON32  encode  156.67 MB/sec
NEON32  decode  117.03 MB/sec
plain   encode  103.30 MB/sec
plain   decode  87.69 MB/sec
Testing with buffer size 1 MB, fastest of 10 * 10
NEON32  encode  156.82 MB/sec
NEON32  decode  117.38 MB/sec
plain   encode  103.32 MB/sec
plain   decode  87.83 MB/sec
Testing with buffer size 100 KB, fastest of 10 * 100
NEON32  encode  180.51 MB/sec
NEON32  decode  128.36 MB/sec
plain   encode  109.30 MB/sec
plain   decode  93.73 MB/sec
Testing with buffer size 10 KB, fastest of 100 * 100
NEON32  encode  182.34 MB/sec
NEON32  decode  129.23 MB/sec
plain   encode  110.00 MB/sec
plain   decode  94.15 MB/sec
Testing with buffer size 1 KB, fastest of 100 * 1000
NEON32  encode  176.32 MB/sec
NEON32  decode  126.49 MB/sec
```

```
Filling buffer with 10.0 MB of random data...
Testing with buffer size 10 MB, fastest of 10 * 1
NEON32  encode  188.33 MB/sec
NEON32  decode  116.03 MB/sec
plain   encode  102.33 MB/sec
plain   decode  86.96 MB/sec
Testing with buffer size 1 MB, fastest of 10 * 10
NEON32  encode  188.29 MB/sec
NEON32  decode  116.21 MB/sec
plain   encode  102.26 MB/sec
plain   decode  87.06 MB/sec
Testing with buffer size 100 KB, fastest of 10 * 100
NEON32  encode  244.81 MB/sec
NEON32  decode  128.04 MB/sec
plain   encode  108.98 MB/sec
plain   decode  93.48 MB/sec
Testing with buffer size 10 KB, fastest of 100 * 100
NEON32  encode  247.87 MB/sec
NEON32  decode  129.04 MB/sec
plain   encode  109.76 MB/sec
plain   decode  93.93 MB/sec
Testing with buffer size 1 KB, fastest of 100 * 1000
NEON32  encode  236.35 MB/sec
NEON32  decode  126.50 MB/sec
plain   encode  108.45 MB/sec
plain   decode  92.85 MB/sec
```