@aklomp 2016-10-17
I tried to compile with Raspbian's stock GCC-4.6.2, but it miscompiles the codec due to a known  bug (long since fixed, but not in Raspbian-land).