@aklomp 2016-10-17
It would be interesting to benchmark the new SIMD encoding stuff on NEON64 against the [native LUT-based codec](https://github.com/aklomp/base64/blob/master/lib/enc/neon64.c) for that arch. The codec (which I'm quite proud of) should be very fast since it uses very few instructions and processes 48 bytes per iteration. However, I was never able to benchmark it because I don't have access to NEON64 hardware (just an emulator: Qemu).