@mayeut 2016-10-17
I did that on the same `iPhone SE`.
Don't look at the decode part, it's been modified (is now faster but I still have some ideas).
The existing encoding for NEON64 is faster than the implementation proposed for NEON32 here (I don't have the figures at hand, but remember what already exists for NEON64 is faster than this on arm64). Well, at least on that specific hardware but given the two implementations uses LUT, that should remain true whatever the hardware (LUT instruction can be quite slow on  some hardware).

```
Filling buffer with 10.0 MB of random data...
Testing with buffer size 10 MB, fastest of 100 * 1
NEON64	encode	4097.03 MB/sec
NEON64	decode	2310.51 MB/sec
plain	encode	1061.21 MB/sec
plain	decode	1235.06 MB/sec
Testing with buffer size 1 MB, fastest of 100 * 10
NEON64	encode	6553.44 MB/sec
NEON64	decode	2311.29 MB/sec
plain	encode	1062.54 MB/sec
plain	decode	1233.68 MB/sec
Testing with buffer size 100 KB, fastest of 100 * 100
NEON64	encode	6533.10 MB/sec
NEON64	decode	2309.77 MB/sec
plain	encode	1063.06 MB/sec
plain	decode	1232.53 MB/sec
Testing with buffer size 10 KB, fastest of 1000 * 100
NEON64	encode	6499.58 MB/sec
NEON64	decode	2307.98 MB/sec
plain	encode	1087.59 MB/sec
plain	decode	1237.40 MB/sec
Testing with buffer size 1 KB, fastest of 1000 * 1000
NEON64	encode	5840.41 MB/sec
NEON64	decode	2273.28 MB/sec
plain	encode	1059.23 MB/sec
plain	decode	1215.20 MB/sec
```