@mayeut 2016-10-17
I pushed the style fixes.